move region back to processes
