+++ hydromt/model/model.py
@@ -36,6 +36,10 @@ from hydromt.plugins import PLUGINS
 
 __all__ = ["Model"]
 
+# define hydromt model entry points
+# see also hydromt.model group in pyproject.toml
+__hydromt_eps__ = ["Model"]
+
 _logger = logging.getLogger(__name__)
 T = TypeVar("T", bound=ModelComponent)
