+++ pyproject.toml
@@ -116,7 +116,7 @@ core = "hydromt.model.components"
 core = "hydromt.data_catalog.drivers"
 
 [project.entry-points."hydromt.models"]
-core = "hydromt.model"
+core = "hydromt.model.model"
 
 [project.entry-points."hydromt.catalogs"]
 core = "hydromt.data_catalog.predefined_catalog"
